UTF8string : Fix + new function.
Fixed a bug in utf8_substr when trying to find an ascii-based text sequence
in a string that contain pure UTF-8 text (non-ascii text) and ascii text.
utf8_len_t was also changed.
Implemented utf8_find to find any UTF-8 text.

+++ src/utf8_types.hpp
@@ -2,6 +2,6 @@
 #define UTF8_TYPES_HPP_INCLUDED
 
 typedef unsigned char byte_t;
-typedef unsigned long utf8_len_t;
+typedef size_t utf8_len_t;
 
 #endif // UTF8_TYPES_HPP_INCLUDED
